[Review] Move llvm.used/llvm.compiler.used helpers from ModuleUtils to GlobalValue; and remove dependency between BitcodeWriter & TransformUtils
